Setup the task creation endpoint
