+++ src/db/mongoose.js
@@ -4,15 +4,4 @@ useNewUrlParser: true,
 useCreateIndex: true
 })
 
-const Task = mongoose.model('Task',{
-    description: {
-        type: String,
-        trim:true,
-        required:true
-    },
-    completed: {
-        type: Boolean,
-        default: false
-    }
-})
