+++ src/index.js
@@ -1,6 +1,7 @@
 const express = require('express')
 require('./db/mongoose')
 const User = require('./models/user')
+const Task = require('./models/task')
 
 const app = express()
 const port = process.env.PORT || 3000
@@ -10,12 +11,23 @@ app.use(express.json())
 app.post('/users',(req, res)=>{
     const user = new User(req.body)
     user.save().then(()=>{
-        res.send(user)
+        res.status(201).send(user)
     }).catch((e)=>{
       res.status(400).send(e)
       
     })
 })
+
+app.post('/tasks',(req , res)=>{
+  const task = new Task(req.body)
+  task.save().then(()=>{
+    res.status().send(task)
+  }).catch((e)=>{
+    res.status(400).send(e)
+  })
+
+})
+
 
 app.listen(port,()=>{
   console.log('the server is up on '+port)
